Take AzureEnvironmentSpecConfig as param to various AgentBaker functions that use it. (#262)
* Take AzureEnvironmentSpecConfig as param to various AgentBaker functions that use it.
The old code used the global mapping from cloud names to their corresponding AzureEnvironmentSpecConfig. It worked because the mapping existed globally in aks-engine code and was used by both AgentBaker and RP. Now AgentBaker and RP both use their own data models, instead of having two copies, we'll keep the canonical copy in RP and pass in the CloudSpecConfigs into AgentBaker functions.

* Remove AzureCloudSpec, AzureGermanCloudSpec, AzureUSGovernmentCloudSpec and AzureChinaCloudSpec.
They are not used in AgentBaker. AgentBaker requires caller to pass in the cloud config spec.

* Use *AzureEnvironmentSpecConfig as param type for assignKubernetesParametersFromAgentProfile and assignKubernetesParameters.
This makes the consistent with other functions in AgentBaker.

* Use *AzureEnvironmentSpecConfig as param everywhere to be consistent.