Remove SetPropertiesDefaults. (#312)
The only usage of SetPropertiesDefaults in AgentBaker is in cmd/generate.go to support "baker generate" call. We've decided to not have any smarts in setting default values in AgentBaker. The logic about defaults used to be in aks-engine code, now for us the logic resides in RP code base. Having AgentBaker maintain a separate defaults settings that may get out of sync with RP is not good for long term code maintenance. Instead, when using AgentBaker, the caller needs to provide the full ContainerService struct, not depending on AgentBaker setting any defaults.
More code can be removed now that this set defaults func is removed. I'll remove those in future PRs.