Remove K8sComponentsByVersionMap from AgentBaker. (#326)
* Remove K8sComponentsByVersionMap from AgentBaker.
AgentBaker now requires user/caller to pass in K8sComponentsByVersionMap. This removes the need to keep AgentBaker's K8sComponentsByVersionMap in sync with its user/caller.

* Pass in only K8sComponents, not the entire K8sComponentsByVersionMap.
We don't need the entire K8sComponentsByVersionMap when generating CSE.

* Define a struct to take exactly what AgentBaker uses from K8sComponents.
This explicitly lists the few strings we need instead of taking in a full map.