Make GetNodeBootstrappingPayload return base64 encoded string. (#412)
* Add a GetNodeBootstrappingPayloadNonARM that provides custom data for non-ARM case.
Updated GetAgentKubernetesLabels and GetAgentKubernetesLabelsDeprecated in the function map to generate the resource group names themselves so they don't need to take it from the nodecustomdata.yml. This simplifies the logic for nodecustomdata.yml: we no longer need to "concat" substrings, which contains a variables('labelResourceGroup') in the middle. We can now directly take nodecustomdata.yaml as a string in our code.

* Remove Windows related change for getting non-ARM custom data.
Windows CSE isn't in AgentBaker yet, so the code there is not worling and isn't worth changing.

* Unify the ...NonARM and ...ARM variant functions that generates custom data.
Because both non-ARM and ARM case use base64 encoding of the custom data, they can be unified into one.

* Remove accidental change to go.mod

* Shouldn't truncate resource group name when adding "z" if its length <= 62.
This is for parity with aks-engine's value:
[if(or(or(endsWith(variables('truncatedResourceGroup'), '-'), endsWith(variables('truncatedResourceGroup'), '_')), endsWith(variables('truncatedResourceGroup'), '.')), concat(take(variables('truncatedResourceGroup'), 62), 'z'), variables('truncatedResourceGroup'))]

I converted that to code incorrectly before.