Pass in the primary scale set name instead of calculating it. (#923)
Currently the primary scale set names are calculated both in RP and AgentBaker. It works because both calculation use the exact same code. But this is fragile because we have duplicated code in two places and they have to be kept in sync. It's better to calculate it in just one place and pass the info in. That way we will not get mismatched names when in the future someone changes the code.