Add an AgentBaker interface for RP to use. (#926)
* Add an AgentBaker interface for RP to use.
This is the preparation step to separate AgentBaker into a standalone service. RP will call the API provided by AgentBaker to get CustomData, CSE, and VHD image to provision nodes.
We'll use protocol buffer and gRPC for the API definition. For now, use straight Go interface as API to help testing the end to end flow.

* Return error if cloud is not found in the OS image settings.

* Remove setting for CoreOS. It's not used.