feat: Log CRI plugin containers on Windows nodes
Add logging of the CRI plugins containers in collect-windows-logs.ps1
via crictl.exe ps -a. Often it's helpful to see the state of the world
from the CRI plugins perspective as well to see if there's any containers
that have EXITED from CRI's perspective but have yet to get removed, or
just to cross reference with other parts of the stack to see any oddities.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>