fix: replace password-related settings properly (#3036)
We have code that modifies /etc/default/useradd /etc/login.defs to bring them into compliance with CIS benchmarks. However, the script does some weird things to those files and in the case of mariner actually fails to set the INACTIVE setting in /etc/useradd.

This change fixes the script to modify settings in place or append them, taking into account the fact that a file may not have a trailing new-line.

I also normalized the tab length in the file, which was different in different functions.