fix: dns timeout to respect resolve.conf 5s timeout and try all servers
add `-timeout=5 -retries=0` to `nslookup` invocation to avoid
default resolv.conf behavior (explicitly reducing retries).

default resolv.conf timeout is 5 sec with 3 retries,
meaning a single nslookup takes 15 sec on the first dns server.

if a user has multiple DNS servers, and the first one is down,
we will never try the 2nd+ one. in theory they could have a
bunch (3+, with 3 respected) but this is a quick fix.

we also appear to do sequential v4/v6 queries with nslookup
and wait for both, which is not really necessary,
but not sure how to easily issue both without some more bash.
additionally, we may have dual stack or ipv6 only apiserver
in the future.