fix: add reboot into correct kernel for mariner
this primarily affects the mariner kata image.
we accidentally remove the kernel in post-install without a reboot.

issue is primarily due to calling enableMarinerKata and then cleaning up kernels somewhat indiscriminately such that we reboot into a non-existing (?) kernel.

we never actually had a reboot even before [we removed one][0],
but the [cleanup for mariner][1] needs one to do the right thing (we hope)

- [enable kata][2]
- [swapping grub cfg][3] (applied to _next_ boot, but no reboot occurs)
- [kernel cleanup][4]

[0]: https://github.com/Azure/AgentBaker/commit/27592e5ebdef0cbd46e0da840ff9a86d57df1d8c
[1]: https://github.com/Azure/AgentBaker/commit/c0612b842b6a060db124793cd29a169386fbc9df
[2]: https://github.com/Azure/AgentBaker/blob/4a7ba42d7abe6b55d069e9c2b3f10da2a958b0b9/vhdbuilder/packer/install-dependencies.sh#L102
[3]: https://github.com/Azure/AgentBaker/blob/4a7ba42d7abe6b55d069e9c2b3f10da2a958b0b9/vhdbuilder/scripts/linux/mariner/tool_installs_mariner.sh#L128-L135
[4]: https://github.com/Azure/AgentBaker/blob/4a7ba42d7abe6b55d069e9c2b3f10da2a958b0b9/vhdbuilder/packer/post-install-dependencies.sh#L36-L39