chore: cache busybox in VHD
In AKS kubernetes version 1.25 and later, kube-proxy init container
runs busybox:1.35. We used to use the kube-proxy image, but had
to switch when upstream started building from distroless.

The busybox image is 4.27 MB. This is downloaded by almost every
AKS node on startup, so I think we should cache it in the VHD.