feat: optimize resize-osdrive to improve node provisioning time
While inspecting node provision time – Resize-OSDrive was taking over 2min in some cases, the Windows Storage team informed us that Get-PartitionSupportedSize can be a slow call as it not only gets the maximum size but also the minimum size which requires a call into defrag. This change removes that call and uses diskpart for the resize operation.