fix: install missing kubenet ebtables drop rule
In kubenet, the script ensure-no-dup.sh is supposed to install
ebtables rules to prevent traffic to service IPs from being
duplicated when the bridge is in promiscuous mode.

The script reads the pod CIDR from the CNI conflist.
However, this broke when the CNI conflist format changed
to support kubenet dual-stack in commit
869183619373769ae24e130afeef3f0dd865c6ef

Since the field `ipam.subnet` no longer exists in the CNI conflist,
the jq expression was returning "null", which was rejected
by ebtables. The result is that when the client and server pod
are on the same node, and the client sends traffic to the service VIP,
the server will see duplicate packets (the same issue described
here: https://github.com/kubernetes/kubernetes/issues/25793)

This commit fixes the issue by updating the jq expression
to read `.ipam.ranges[0][0].subnet` instead.

This addresses the immediate issue for IPv4, but longer-term we need to:
1. Test whether IPv6 is also seeing this issue (likely yes)
2. Consider configuring the bridge plugin to use hairpinMode=true,
   promiscMode=false and removing the ensure-no-dup script entirely.