e2e: Rename custom region test and simplify config
The Test_MyScenario_BasicDeployment function has been renamed to
Test_AlternateRegion_BasicDeployment to better reflect its purpose of testing
deployments in a non-default region.

The VMConfigMutator that explicitly set the location on the VMSS has been
removed as this is now correctly handled by the Scenario.Location field,
simplifying the test configuration.

Signed-off-by: Suraj Deshmukh <suraj.deshmukh@microsoft.com>