e2e: Centralize blob storage in default location
This commit refactors the Azure client code to ensure that the blob storage
account and container are always created in the `DefaultLocation`, regardless of
the location of the VM identity being created.

Previously, the code attempted to create these resources in the same location as
the VM identity, which is incorrect as the storage is intended to be
centralized.

The following changes have been made:
- The `createBlobStorageAccount` and `createBlobStorageContainer` functions now
  consistently use `Config.DefaultLocation`.
- The `location` parameter in `CreateVMManagedIdentity` has been renamed to
  `identityLocation` for improved clarity.
- Comments have been updated to reflect this centralized storage strategy.

Signed-off-by: Suraj Deshmukh <suraj.deshmukh@microsoft.com>