e2e: Add flag to control Nvidia device plugin installation
This change introduces a mechanism to control the installation of the Nvidia
device plugin during the E2E cluster preparation phase. This is necessary for
scenarios, such as testing the GPU operator, where the device plugin should not
be pre-installed by the test framework.

The `prepareCluster` function now accepts an `installNvidiaDevicePlugin` boolean
flag, which is passed down to `EnsureDebugDaemonsets` to conditionally deploy
the `nvidia-device-plugin` daemonset.

A new cluster factory, `ClusterKubenetNoNvidiaDevicePlugin`, has been added to
create clusters without the plugin. The `Test_Ubuntu2404_GPU_H100` test is
updated to use this new cluster type, and the now-redundant
`EnableGPUDevicePluginIfNeeded` bootstrap setting has been removed from the test
configuration.

Signed-off-by: Suraj Deshmukh <suraj.deshmukh@microsoft.com>