test(e2e): Fix race when we need to replicate an image to two regions
VHDUbuntu2404Gen2Containerd may need to be replicated to two regions at the
same time: southafricanorth and westus3.  With how we're currently executing
the replications, both will start running at the same time, but both supply the
entire region replica list so one will overwrite the other.

This is what it looks like in the log:

   ##[warning]Replicating to region westus3, available regions: West US 2, image version /subscriptions/c4c3550e-a965-4993-a50c-628fd38cd3e1/resourceGroups/aksvhdtestbuildrg/providers/Microsoft.Compute/galleries/PackerSigGalleryEastUS/images/2404gen2containerd/versions/1.1753949350.11431
   ##[warning]Replicating to region southafricanorth, available regions: West US 2, image version /subscriptions/c4c3550e-a965-4993-a50c-628fd38cd3e1/resourceGroups/aksvhdtestbuildrg/providers/Microsoft.Compute/galleries/PackerSigGalleryEastUS/images/2404gen2containerd/versions/1.1753949350.11431

Fix the race by serializing the start of the replication and combine it with a
lookup of the latest image list. This shouldn't be heavily contested because
the replication itself is what takes long.