feat: Make aks-node-controller.service work on Flatcar
Flatcar doesn't ship with cloud-init, it relies on ignition, but there is also
a fallback tool: coreos-cloudinit. It can support a subset of cloudinit
directives. Since the customdata for scriptless is really short, we're going to
try to make scriptless work with coreos-cloudinit for Flatcar.

To make that work we need two changes to the aks-node-controller.service. To
get the unit to start we need to add WantedBy=multi-user.target as a second
option (could be the only one too), because Flatcar doesn't have a
cloud-init.target. The other is ordering: we need to make sure
aks-node-controller.service runs after coreos-cloudinit, which runs as
oem-cloudinit.service. However oem-cloudinit.service is not enabled by default,
it is enabled at runtime through enable-oem-cloudinit.service. So also need to
order on that, otherwise systemd will drop the ordering constraint on
oem-cloudinit.service (since it is not queued up).

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>